refactor: fix typos and change code structure
