+++ src/common/function/src/function_registry.rs
@@ -18,13 +18,13 @@ use std::sync::{Arc, RwLock};
 
 use once_cell::sync::Lazy;
 
+use crate::function::FunctionRef;
 use crate::scalars::aggregate::{AggregateFunctionMetaRef, AggregateFunctions};
 use crate::scalars::date::DateFunction;
-use crate::scalars::function::FunctionRef;
 use crate::scalars::math::MathFunction;
 use crate::scalars::numpy::NumpyFunction;
 use crate::scalars::timestamp::TimestampFunction;
-use crate::system::build::SystemFunction;
+use crate::system::SystemFunction;
 
 #[derive(Default)]
 pub struct FunctionRegistry {
