+++ src/common/function/src/lib.rs
@@ -15,4 +15,6 @@
 pub mod scalars;
 pub mod system;
 
+pub mod function;
+pub mod function_registry;
 pub mod helper;
