+++ src/common/function/src/scalars/numpy.rs
@@ -19,7 +19,7 @@ use std::sync::Arc;
 
 use clip::ClipFunction;
 
-use crate::scalars::function_registry::FunctionRegistry;
+use crate::function_registry::FunctionRegistry;
 
 pub(crate) struct NumpyFunction;
