+++ src/common/function/src/system.rs
@@ -13,3 +13,17 @@
 // limitations under the License.
 
 pub mod build;
+
+use std::sync::Arc;
+
+use build::BuildFunction;
+
+use crate::function_registry::FunctionRegistry;
+
+pub(crate) struct SystemFunction;
+
+impl SystemFunction {
+    pub fn register(registry: &FunctionRegistry) {
+        registry.register(Arc::new(BuildFunction));
+    }
+}
