+++ src/query/src/query_engine.rs
@@ -22,8 +22,9 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use catalog::CatalogManagerRef;
 use common_base::Plugins;
+use common_function::function::FunctionRef;
+use common_function::function_registry::FUNCTION_REGISTRY;
 use common_function::scalars::aggregate::AggregateFunctionMetaRef;
-use common_function::scalars::{FunctionRef, FUNCTION_REGISTRY};
 use common_query::prelude::ScalarUdf;
 use common_query::Output;
 use datatypes::schema::Schema;
